+++ app/src/main/java/com/survivalcoding/network_apps/feature_basic/di/TodoDataSourceModule.kt
@@ -1,4 +1,4 @@
-package com.survivalcoding.network_apps.feature_basic.data.todo
+package com.survivalcoding.network_apps.feature_basic.di
 
 import com.survivalcoding.network_apps.feature_basic.data.todo.datasource.TodoDataSource
 import com.survivalcoding.network_apps.feature_basic.data.todo.datasource.local.LocalTodoDataSource
